Fix UB: signed overflow in AES-256 key loading shifts
The key-schedule initialisation widens key bytes with `key[4 * i] << 24`.
The bytes are uint8_t, so they are promoted to int before the shift; any
key byte >= 0x80 shifted into the sign bit is undefined behaviour
(e.g. `140 << 24`). UBSAN reports this on both code paths:

  aes256.c:368:24: runtime error: left shift of 140 by 24 places cannot be
  represented in type 'int'

Because MTProto keys are random, roughly half of all key bytes are >= 0x80,
so the UB is exercised on essentially every call to the software key
schedule - including on AES-NI builds, which called it too before the
key-schedule change. GCC/Clang currently emit the expected
two's-complement result, so no ciphertext corruption was observed, but the
expression is not defined by the standard and the compiler is entitled to
assume it never occurs.

Cast each byte to uint32_t before shifting, as already done in sub_word()
and in the aes256_encrypt/aes256_decrypt SBOX1 terms.

Validated: output is bit-for-bit identical to the reference build (63-case
differential plus the full 3886-test suite on both an AES-NI and a
TGCRYPTO_NO_AESNI build), and ASAN/UBSAN is now clean on the software path,
where it previously reported the error above.

+++ tgcrypto/aes256.c
@@ -365,10 +365,13 @@ void aes256_set_encryption_key(const uint8_t key[32], uint32_t expandedKey[60]) 
 
     for (i = 0; i < Nk; ++i)
         expandedKey[i] = (
-            key[4 * i] << 24
-            | key[4 * i + 1] << 16
-            | key[4 * i + 2] << 8
-            | key[4 * i + 3]
+            /* Each byte is widened to uint32_t before shifting: shifting a
+             * promoted int that overflows the sign bit (e.g. 140 << 24) is
+             * undefined behaviour. */
+            (uint32_t) key[4 * i] << 24
+            | (uint32_t) key[4 * i + 1] << 16
+            | (uint32_t) key[4 * i + 2] << 8
+            | (uint32_t) key[4 * i + 3]
         );
 
     for (i = Nk; i < Nb * (Nr + 1); ++i) {
